requirements.txt
google-genai>=0.3.0
opencv-python>=4.8.0
